feat(autoware_lidar_centerpoint): Add intensity model  (#2123)
* Add intensity points

* Add intensity points

* Add intensity points

* Add intensity points

* Add intensity points

* Add intensity points

* Speed up memory access

* Remove unused code

* style(pre-commit): autofix

---------

+++ perception/autoware_lidar_centerpoint/include/autoware/lidar_centerpoint/preprocess/pointcloud_densification.hpp
@@ -50,7 +50,7 @@ private:
 
 struct PointCloudWithTransform
 {
-  cuda::unique_ptr<float[]> points_d{nullptr};
+  cuda::unique_ptr<uint8_t[]> points_d{nullptr};
   std_msgs::msg::Header header;
   std::size_t num_points{0};
   std::size_t point_step{0};
